+++ CHANGELOG.md
@@ -9,6 +9,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Sync to upstream Luau 0.610
+- Non-function properties will now no longer be shown by default when autocompleting a method call (e.g., `foo:bar`).
+  To revert back to the original behaviour, enable `luau-lsp.completion.showPropertiesOnMethodCall`
 
 ## [1.27.1] - 2024-01-20